Further optimize stage1, restricting stage2 size to 32MiB
This saves a further 16 bytes of code by treating the LBA address
as a 16-bit integer (hence the 32MiB restriction - srcfs is not
included in this, as it's read by stage2 using full 32-bit LBA),
and by not passing the LBA address around in the code, but only
storing it in the address packet.

From the data section, 2 bytes are saved by embedding the last read
offset in the address packet, in the space normally occupied by the
number of sectors to read. This is possible because sector reads
always happen when the offset is 0x1FF, the number of sectors is
then unconditionally rewritten to 1, and after the read, the offset
is always reset to 0. Hence the same 2 bytes can be reused for both
functions.

With this, stage1 consists of just 176 bytes of code, and 16 bytes
of initialized data, excluding MBR boilerplate.